+++ packages/common/ProForm/src/interactions/useForm.ts
@@ -8,6 +8,7 @@ export function useForm(formCustomization: FormCustomization): UseForm {
 		{
 			submit: formCustomizer.submit.bind(formCustomizer),
 			hydrate: formCustomizer.hydrate.bind(formCustomizer),
+			share: formCustomizer.share.bind(formCustomizer),
 		},
 	];
 }