+++ packages/common/ProForm/src/types/useFormTypes.ts
@@ -8,5 +8,6 @@ export type UseForm = [
 	{
 		submit: () => Promise<AnyObject>;
 		hydrate: (data: AnyObject) => any;
+		share: (data: AnyObject) => any;
 	},
 ];